+++ combat.h
@@ -10,7 +10,7 @@
 #include "characterCode/Player.h"
 
 void combat();
-float typeRacer(Enemy& enemy,
+double typeRacer(Enemy& enemy,
     int& wordsWritten, int& lettersWritten, int& wrongLetters,
     double& damage, double& accuracyPercent, double& wordsPerMinute,
     Player& player);
